refactor: add domain and operations
