+++ src/domain/entities/navbar/NavItemType.ts
